+++ Aban360.LocationPool.Domain/Features/MainHierarchy/Zone.cs
@@ -5,11 +5,11 @@ namespace Aban360.LocationPool.Domain.Features.MainHierarchy;
 [Table(nameof(Zone))]
 public class Zone
 {
-    public short Id { get; set; }
+    public int Id { get; set; }
 
     public string Title { get; set; } = null!;
 
-    public short RegionId { get; set; }
+    public int RegionId { get; set; }
 
     public string UnstandardCode { get; set; } = null!;
